% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{Regr2STest package}
\alias{Regr2STest package}
\alias{Regr2STest package-package}
\title{Regression Two Sample Test}
\description{
A package to compute regression two sample tesst using 
probabilistic classifiers.
}
