% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_sample_test.R
\docType{data}
\name{TwoSampleRegressionTest}
\alias{TwoSampleRegressionTest}
\title{TwoSampleRegressionTest}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
TwoSampleRegressionTest
}
\description{
The TwoSampleRegressionTest class. Provides functions for computing the
two sample regression test
}
\section{Usage}{

\preformatted{model <- NNKCDE$new(classifier_name)
model$ts_test(first_sample, second_sample, n_bootstrap)
}
}

\section{Arguments}{

\code{classifier_name} Name of the classifier used in the regression test.
The availbility of classifiers is dependent on what it has been implemented.

\code{first_sample} First sample of the two sample testing test

\code{second_sample} Second sample of the two sample testing test

\code{n_bootstrap} Number of boostrap repetitions for calculating the p-value
of the two sample test.
}

\section{Methods}{

\code{$new(classifier_name)} Initializes a new TwoSampleRegressionTest
 object.

\code{$regstat(X, y} function to calculate the test statistic used in the
test. X is the matrix of covariates, while y is a boolean vector indicating
which sample the datum belongs to (either first or second sample)

\code{$ts_test(first_sample, second_sample, n_bootstrap)} Calculates the 
p-value for the two sample regression test.
}

\examples{
\dontrun{
model <- TwoSampleRegressionTest$new('Log Regr.')
model$ts_test(rnorm(500), rnorm(500))
}
}
\keyword{datasets}
